complex_one/pyproject.toml
[project]
name = "complex-one"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firecrawl>=2.16.2",
    "langchain>=0.3.26",
    "langchain-groq>=0.3.6",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.5.4",
    "mcp>=1.12.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "typing>=3.10.0.0",
]
